Remove app authentication e suas dependências, redirecionando URLs para o gerenciamento de users.
