--- autismcare/authentication/serializers/__init__.py
@@ -1,10 +1,0 @@
-from .base_serializer import AutenticationUserSerializer, AddressSerializer
-from .role_serializer import RoleSerializer
-from .employee_serializer import EmployeeSerializer
-from .parent_serializer import ParentSerializer
-from .child_serializer import ChildSerializer
-from .medical_serializer import (
-    AllergySerializer,
-    MedicalRecordSerializer,
-    MedicalHistorySerializer,
-)
