--- autismcare/authentication/views/__init__.py
@@ -1,6 +1,0 @@
-from .base_view import AutenticationUserViewSet, AddressViewSet
-from .role_view import RoleViewSet
-from .employee_view import EmployeeViewSet
-from .parent_view import ParentViewSet
-from .child_view import ChildViewSet
-from .medical_view import AllergyViewSet, MedicalRecordViewSet, MedicalHistoryViewSet
