+++ autismcare/users/permissions.py
@@ -1,0 +1,32 @@
+from rest_framework import permissions
+
+class IsOwnerOrReadOnly(permissions.BasePermission):
+    """
+    Permissão personalizada para permitir que apenas os donos de um objeto o editem.
+    """
+    def has_object_permission(self, request, view, obj):
+        if request.method in permissions.SAFE_METHODS:
+            return True
+        return obj.user == request.user
+
+class IsParentOrReadOnly(permissions.BasePermission):
+    """
+    Permissão que verifica se o utilizador logado é o pai/mãe do objeto (Child)
+    ou um funcionário.
+    """
+    def has_object_permission(self, request, view, obj):
+        if hasattr(request.user, 'employee'):
+            return True
+        
+        if request.method in permissions.SAFE_METHODS:
+            return True
+        return obj.parent.user == request.user
+
+class IsEmployeeOrReadOnly(permissions.BasePermission):
+    """
+    Permite acesso total para funcionários (employees) e apenas leitura para outros.
+    """
+    def has_permission(self, request, view):
+        if request.method in permissions.SAFE_METHODS:
+            return True
+        return hasattr(request.user, 'employee')
